chore: sampling improvements
