+++ pkg/storer/reserve.go
@@ -5,28 +5,17 @@
 package storer
 
 import (
-	"bytes"
 	"context"
-	"crypto/hmac"
-	"encoding/binary"
 	"encoding/hex"
 	"errors"
 	"fmt"
 	"math/big"
-	"sort"
 	"sync"
-	"testing"
 	"time"
 
-	"github.com/ethersphere/bee/pkg/bmt"
 	"github.com/ethersphere/bee/pkg/postage"
-	"github.com/ethersphere/bee/pkg/soc"
 	storage "github.com/ethersphere/bee/pkg/storage"
-	chunk "github.com/ethersphere/bee/pkg/storage/testing"
-	"github.com/ethersphere/bee/pkg/storer/internal/reserve"
 	"github.com/ethersphere/bee/pkg/swarm"
-	"go.uber.org/atomic"
-	"golang.org/x/sync/errgroup"
 )
 
 const (
@@ -385,316 +374,6 @@ func (db *DB) SubscribeBin(ctx context.Context, bin uint8, start uint64) (<-chan
 	}, errC
 }
 
-type SampleItem struct {
-	TransformedAddress swarm.Address
-	ChunkAddress       swarm.Address
-	ChunkData          []byte
-	Stamp              *postage.Stamp
-}
-
-func newStamp(s swarm.Stamp) *postage.Stamp {
-	return postage.NewStamp(s.BatchID(), s.Index(), s.Timestamp(), s.Sig())
-}
-
-type Sample struct {
-	Items []SampleItem
-}
-
-func RandSample(t *testing.T, anchor []byte) Sample {
-	t.Helper()
-
-	hasher := bmt.NewTrHasher(anchor)
-
-	items := make([]SampleItem, SampleSize)
-	for i := 0; i < SampleSize; i++ {
-		ch := chunk.GenerateTestRandomChunk()
-
-		tr, err := transformedAddress(hasher, ch, swarm.ChunkTypeContentAddressed)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		items[i] = SampleItem{
-			TransformedAddress: tr,
-			ChunkAddress:       ch.Address(),
-			ChunkData:          ch.Data(),
-			Stamp:              newStamp(ch.Stamp()),
-		}
-	}
-
-	sort.Slice(items, func(i, j int) bool {
-		return items[i].TransformedAddress.Compare(items[j].TransformedAddress) == -1
-	})
-
-	return Sample{Items: items}
-}
-
-// ReserveSample generates the sample of reserve storage of a node required for the
-// storage incentives agent to participate in the lottery round. In order to generate
-// this sample we need to iterate through all the chunks in the node's reserve and
-// calculate the transformed hashes of all the chunks using the anchor as the salt.
-// In order to generate the transformed hashes, we will use the std hmac keyed-hash
-// implementation by using the anchor as the key. Nodes need to calculate the sample
-// in the most optimal way and there are time restrictions. The lottery round is a
-// time based round, so nodes participating in the round need to perform this
-// calculation within the round limits.
-// In order to optimize this we use a simple pipeline pattern:
-// Iterate chunk addresses -> Get the chunk data and calculate transformed hash -> Assemble the sample
-func (db *DB) ReserveSample(
-	ctx context.Context,
-	anchor []byte,
-	storageRadius uint8,
-	consensusTime uint64,
-	minBatchBalance *big.Int,
-) (Sample, error) {
-
-	g, ctx := errgroup.WithContext(ctx)
-
-	chunkC := make(chan reserve.ChunkItem)
-	var stat sampleStat
-
-	excludedBatchIDs, err := db.batchesBelowValue(minBatchBalance)
-	if err != nil {
-		db.logger.Error(err, "get batches below value")
-	}
-
-	t := time.Now()
-
-	// Phase 1: Iterate chunk addresses
-	g.Go(func() error {
-		defer close(chunkC)
-		iterationStart := time.Now()
-
-		err := db.reserve.IterateChunksItems(db.repo, storageRadius, func(chi reserve.ChunkItem) (bool, error) {
-			select {
-			case chunkC <- chi:
-				stat.TotalIterated.Inc()
-				return false, nil
-			case <-ctx.Done():
-				return false, ctx.Err()
-			}
-		})
-		if err != nil {
-			return err
-		}
-
-		stat.IterationDuration.Add(time.Since(iterationStart).Nanoseconds())
-		return nil
-	})
-
-	// Phase 2: Get the chunk data and calculate transformed hash
-	sampleItemChan := make(chan SampleItem)
-	const workers = 6
-
-	for i := 0; i < workers; i++ {
-		g.Go(func() error {
-			// PH4_Logic:
-			// hasher := bmt.NewTrHasher(anchor)
-
-			hmacr := hmac.New(swarm.NewHasher, anchor)
-
-			for chItem := range chunkC {
-				getStart := time.Now()
-
-				stat.GetDuration.Add(time.Since(getStart).Nanoseconds())
-
-				// exclude chunks who's batches balance are below minimum
-				if _, found := excludedBatchIDs[string(chItem.Chunk.Stamp().BatchID())]; found {
-					stat.BelowBalanceIgnored.Inc()
-					continue
-				}
-
-				// check if the timestamp on the postage stamp is not later than
-				// the consensus time.
-				if binary.BigEndian.Uint64(chItem.Chunk.Stamp().Timestamp()) > consensusTime {
-					stat.NewIgnored.Inc()
-					continue
-				}
-
-				// Skip chunks if they are not SOC or CAC
-				if chItem.Type != swarm.ChunkTypeContentAddressed && chItem.Type != swarm.ChunkTypeSingleOwner {
-					continue
-				}
-
-				hmacrStart := time.Now()
-
-				hmacr.Reset()
-				_, err := hmacr.Write(chItem.Chunk.Data())
-				if err != nil {
-					return err
-				}
-				taddr := swarm.NewAddress(hmacr.Sum(nil))
-
-				// PH4_Logic::
-				// taddr, err := transformedAddress(hasher, chItem.Chunk, chItem.Type)
-				// if err != nil {
-				// 	return err
-				// }
-
-				stat.HmacrDuration.Add(time.Since(hmacrStart).Nanoseconds())
-
-				select {
-				case sampleItemChan <- SampleItem{
-					TransformedAddress: taddr,
-					ChunkAddress:       chItem.Chunk.Address(),
-					ChunkData:          chItem.Chunk.Data(),
-					Stamp:              newStamp(chItem.Chunk.Stamp()),
-				}:
-				case <-ctx.Done():
-					return ctx.Err()
-				}
-			}
-
-			return nil
-		})
-	}
-
-	go func() {
-		_ = g.Wait()
-		close(sampleItemChan)
-	}()
-
-	sampleItems := make([]SampleItem, 0, SampleSize)
-	// insert function will insert the new item in its correct place. If the sample
-	// size goes beyond what we need we omit the last item.
-	insert := func(item SampleItem) {
-		added := false
-		for i, sItem := range sampleItems {
-			if le(item.TransformedAddress, sItem.TransformedAddress) {
-				sampleItems = append(sampleItems[:i+1], sampleItems[i:]...)
-				sampleItems[i] = item
-				added = true
-				break
-			}
-		}
-		if len(sampleItems) > SampleSize {
-			sampleItems = sampleItems[:SampleSize]
-		}
-		if len(sampleItems) < SampleSize && !added {
-			sampleItems = append(sampleItems, item)
-		}
-	}
-
-	// Phase 3: Assemble the sample. Here we need to assemble only the first SampleSize
-	// no of items from the results of the 2nd phase.
-	for item := range sampleItemChan {
-		currentMaxAddr := swarm.EmptyAddress
-		if len(sampleItems) > 0 {
-			currentMaxAddr = sampleItems[len(sampleItems)-1].TransformedAddress
-		}
-
-		if le(item.TransformedAddress, currentMaxAddr) || len(sampleItems) < SampleSize {
-			ch := swarm.NewChunk(item.ChunkAddress, item.ChunkData).WithStamp(item.Stamp)
-			if _, err := db.validStamp(ch); err != nil {
-				stat.InvalidStamp.Inc()
-				db.logger.Debug("invalid stamp for chunk", "chunk_address", ch.Address(), "error", err)
-				continue
-			}
-
-			insert(item)
-		}
-	}
-
-	if err := g.Wait(); err != nil {
-		return Sample{}, fmt.Errorf("sampler: failed creating sample: %w", err)
-	}
-
-	sample := Sample{
-		Items: sampleItems,
-	}
-
-	db.logger.Info("reserve sampler done", "duration", time.Since(t), "storage_radius", storageRadius, "consensus_time_ns", consensusTime, "stats", stat, "sample", sample)
-
-	return sample, nil
-}
-
 func (db *DB) po(addr swarm.Address) uint8 {
 	return swarm.Proximity(db.baseAddr.Bytes(), addr.Bytes())
-}
-
-// less function uses the byte compare to check for lexicographic ordering
-func le(a, b swarm.Address) bool {
-	return bytes.Compare(a.Bytes(), b.Bytes()) == -1
-}
-
-func transformedAddress(hasher *bmt.Hasher, chunk swarm.Chunk, chType swarm.ChunkType) (swarm.Address, error) {
-	switch chType {
-	case swarm.ChunkTypeContentAddressed:
-		return transformedAddressCAC(hasher, chunk)
-	case swarm.ChunkTypeSingleOwner:
-		return transformedAddressSOC(hasher, chunk)
-	default:
-		return swarm.ZeroAddress, fmt.Errorf("chunk type [%v] is is not valid", chType)
-	}
-}
-
-func transformedAddressCAC(hasher *bmt.Hasher, chunk swarm.Chunk) (swarm.Address, error) {
-	hasher.Reset()
-	hasher.SetHeader(chunk.Data()[:bmt.SpanSize])
-
-	_, err := hasher.Write(chunk.Data()[bmt.SpanSize:])
-	if err != nil {
-		return swarm.ZeroAddress, err
-	}
-
-	taddr, err := hasher.Hash(nil)
-	if err != nil {
-		return swarm.ZeroAddress, err
-	}
-
-	return swarm.NewAddress(taddr), nil
-}
-
-func transformedAddressSOC(hasher *bmt.Hasher, chunk swarm.Chunk) (swarm.Address, error) {
-	// Calculate transformed address from wrapped chunk
-	sChunk, err := soc.FromChunk(chunk)
-	if err != nil {
-		return swarm.ZeroAddress, err
-	}
-	taddrCac, err := transformedAddressCAC(hasher, sChunk.WrappedChunk())
-	if err != nil {
-		return swarm.ZeroAddress, err
-	}
-
-	// Hash address and transformed address to make transformed address for this SOC
-	sHasher := swarm.NewHasher()
-	if _, err := sHasher.Write(chunk.Address().Bytes()); err != nil {
-		return swarm.ZeroAddress, err
-	}
-	if _, err := sHasher.Write(taddrCac.Bytes()); err != nil {
-		return swarm.ZeroAddress, err
-	}
-
-	return swarm.NewAddress(sHasher.Sum(nil)), nil
-}
-
-type sampleStat struct {
-	TotalIterated       atomic.Int64
-	NotFound            atomic.Int64
-	NewIgnored          atomic.Int64
-	InvalidStamp        atomic.Int64
-	BelowBalanceIgnored atomic.Int64
-	IterationDuration   atomic.Int64
-	GetDuration         atomic.Int64
-	HmacrDuration       atomic.Int64
-	ValidStampDuration  atomic.Int64
-}
-
-func (s sampleStat) String() string {
-
-	seconds := int64(time.Second)
-
-	return fmt.Sprintf(
-		"Chunks: %d NotFound: %d New Ignored: %d InvalidStamp: %d BelowBalanceIgnored: %d Iteration Duration: %d secs GetDuration: %d secs"+
-			" HmacrDuration: %d secs ValidStampDuration: %d secs",
-		s.TotalIterated.Load(),
-		s.NotFound.Load(),
-		s.NewIgnored.Load(),
-		s.InvalidStamp.Load(),
-		s.BelowBalanceIgnored.Load(),
-		s.IterationDuration.Load()/seconds,
-		s.GetDuration.Load()/seconds,
-		s.HmacrDuration.Load()/seconds,
-		s.ValidStampDuration.Load()/seconds,
-	)
 }
